+++ src/views/admin-configuration/experience-config.vue
@@ -120,30 +120,46 @@ export default {
      * @description 删除经验配置
      */
     deleteData (cardItem) {
-      this.$store.dispatch('setting/showLoading', this.routeName)
-      const reqData = { staffCardId: cardItem.id }
-      OperationManage.deleteCard(reqData)
-        .then(() => {
+      this.$confirm('确认删除该经验配置吗？', '', {
+        confirmButtonText: '确认',
+        cancelButtonText: '取消',
+        type: 'warning',
+        center: true
+      }).then(async () => {
+        try {
+          this.$store.dispatch('setting/showLoading', this.routeName)
+          const reqData = { staffCardId: cardItem.id }
+          await OperationManage.deleteCard(reqData)
           this.$newMessage.success('删除成功')
           this.getExperienceConfigList()
-        })
-      this.$store.dispatch('setting/hiddenLoading', this.routeName)
+        } catch (error) {
+          this.$store.dispatch('setting/hiddenLoading', this.routeName)
+          console.error(error)
+        }
+      }).catch(() => {})
     },
     /**
      * @description 提前结束
      */
     async closeCard (cardItem) {
-      try {
-        this.$store.dispatch('setting/showLoading', this.routeName)
-        const reqData = { staffCardId: cardItem.id }
-        await OperationManage.closeCard(reqData)
-        this.$newMessage.success('提前结束')
-        this.getExperienceConfigList()
-        this.$store.dispatch('setting/hiddenLoading', this.routeName)
-      } catch (error) {
-        this.$store.dispatch('setting/hiddenLoading', this.routeName)
-        console.error(error)
-      }
+      this.$confirm('确认提前结束吗？', '', {
+        confirmButtonText: '确认',
+        cancelButtonText: '取消',
+        type: 'warning',
+        center: true
+      }).then(async () => {
+        try {
+          this.$store.dispatch('setting/showLoading', this.routeName)
+          const reqData = { staffCardId: cardItem.id }
+          await OperationManage.closeCard(reqData)
+          this.$newMessage.success('提前结束')
+          this.getExperienceConfigList()
+          this.$store.dispatch('setting/hiddenLoading', this.routeName)
+        } catch (error) {
+          this.$store.dispatch('setting/hiddenLoading', this.routeName)
+          console.error(error)
+        }
+      }).catch(() => {})
     },
     /**
      * @description 获取经验配置列表
