+++ src/views/admin-configuration/gold-config.vue
@@ -116,33 +116,47 @@ export default {
      * @description 删除金币配置
      */
     async deleteData (cardItem) {
-      try {
-        this.$store.dispatch('setting/showLoading', this.routeName)
-        const reqData = { staffCardId: cardItem.id }
-        await OperationManage.deleteCard(reqData)
-        this.$newMessage.success('删除成功')
-        this.getGoldConfigList()
-        this.$store.dispatch('setting/hiddenLoading', this.routeName)
-      } catch (error) {
-        this.$store.dispatch('setting/hiddenLoading', this.routeName)
-        console.error(error)
-      }
+      this.$confirm('确认删除该金币奖励吗？', '', {
+        confirmButtonText: '确认',
+        cancelButtonText: '取消',
+        type: 'warning',
+        center: true
+      }).then(async () => {
+        try {
+          this.$store.dispatch('setting/showLoading', this.routeName)
+          const reqData = { staffCardId: cardItem.id }
+          await OperationManage.deleteCard(reqData)
+          this.$newMessage.success('删除成功')
+          this.getGoldConfigList()
+          this.$store.dispatch('setting/hiddenLoading', this.routeName)
+        } catch (error) {
+          this.$store.dispatch('setting/hiddenLoading', this.routeName)
+          console.error(error)
+        }
+      }).catch(() => {})
     },
     /**
      * @description 提前结束
      */
     async closeCard (cardItem) {
-      try {
-        this.$store.dispatch('setting/showLoading', this.routeName)
-        const reqData = { staffCardId: cardItem.id }
-        await OperationManage.closeCard(reqData)
-        this.$newMessage.success('提前结束')
-        this.getGoldConfigList()
-        this.$store.dispatch('setting/hiddenLoading', this.routeName)
-      } catch (error) {
-        this.$store.dispatch('setting/hiddenLoading', this.routeName)
-        console.error(error)
-      }
+      this.$confirm('确认提前结束吗？', '', {
+        confirmButtonText: '确认',
+        cancelButtonText: '取消',
+        type: 'warning',
+        center: true
+      }).then(async () => {
+        try {
+          this.$store.dispatch('setting/showLoading', this.routeName)
+          const reqData = { staffCardId: cardItem.id }
+          await OperationManage.closeCard(reqData)
+          this.$newMessage.success('提前结束')
+          this.getGoldConfigList()
+          this.$store.dispatch('setting/hiddenLoading', this.routeName)
+        } catch (error) {
+          this.$store.dispatch('setting/hiddenLoading', this.routeName)
+          console.error(error)
+        }
+      }).catch(() => {})
     },
     /**
      * @description 获取金币配置列表
